+++ src/test/resources/schema.sql
@@ -4,7 +4,7 @@ CREATE TABLE IF NOT EXISTS product
     product_name       VARCHAR(128) NOT NULL,
     category           VARCHAR(32)  NOT NULL,
     image_url          VARCHAR(256) NOT NULL,
-    price              INT          NOT NULL,
+    price              DECIMAL      NOT NULL,
     stock              INT          NOT NULL,
     description        VARCHAR(1024),
     created_date       TIMESTAMP    NOT NULL,
@@ -24,16 +24,16 @@ CREATE TABLE IF NOT EXISTS `order`
 (
     order_id           INT       NOT NULL PRIMARY KEY AUTO_INCREMENT,
     user_id            INT       NOT NULL,
-    total_amount       INT       NOT NULL, -- 訂單總花費
+    total_amount       DECIMAL   NOT NULL, -- 訂單總花費
     created_date       TIMESTAMP NOT NULL,
     last_modified_date TIMESTAMP NOT NULL
 );
 
 CREATE TABLE IF NOT EXISTS order_item
 (
-    order_item_id INT NOT NULL PRIMARY KEY AUTO_INCREMENT,
-    order_id      INT NOT NULL,
-    product_id    INT NOT NULL,
-    quantity      INT NOT NULL, -- 商品數量
-    amount        INT NOT NULL  -- 商品花費
+    order_item_id INT     NOT NULL PRIMARY KEY AUTO_INCREMENT,
+    order_id      INT     NOT NULL,
+    product_id    INT     NOT NULL,
+    quantity      INT     NOT NULL, -- 商品數量
+    amount        DECIMAL NOT NULL  -- 商品花費
 );
